requirements.txt
streamlit==1.32.0
opencv-python-headless==4.7.0.72
numpy==1.26.0
Pillow==10.0.0
tensorflow==2.14.0
keras==2.14.0
